chore(generic-binding): move to addons package
